dify-plugin-data-analyzer/requirements.txt
# Dify Plugin SDK
dify_plugin>=0.4.0,<0.7.0

# Core dependencies
requests
python-dotenv

# LangGraph for workflow orchestration (1.0.0+)
langgraph>=1.0.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Data science libraries (used by generated code during execution)
numpy
pandas
tables  # HDF5 support for pandas (PyTables, required to avoid dependency warnings)
scikit-learn
statsmodels  # Statistical modeling library (used by LLM-generated code)
seaborn
matplotlib
openpyxl  # Excel file processing

# OpenAI SDK for LLM API calls
openai
